+++ RestaurantManagement.Application/Features/BillFeature/Queries/GetBillById/GetBillByIdQueryHandler.cs
@@ -39,16 +39,16 @@ public class GetBillByIdQueryHandler(IApplicationDbContext context) : IQueryHand
                 a.Booking.BookingDate,
                 a.Booking.BookingTime,
                 a.Order.OrderId,
-                a.Total,
+                (int)a.Total,
                 a.PaymentType,
                 a.Order.OrderDetails.Select(b => new OrderDetailResponse(
                     b.OrderDetailId,
                     b.MealId,
                     b.Meal.MealName,
-                    b.Meal.Price,
+                    (int)b.Meal.Price,
                     b.Meal.ImageUrl,
                     b.Quantity,
-                    b.UnitPrice
+                    (int)b.UnitPrice
                 )).ToArray()))
                 .FirstOrDefaultAsync();
